Creation of new module for inelasticities.
